move:  게임컴포넌트 위치 변경
